backend/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
flask-restx==0.5.1
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-dialogflow==2.9.1
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.3
packaging==21.0
pandas==1.3.4
passlib==1.7.4
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.1
google-cloud-dialogflow==2.9.1
pymessenger==0.0.7.0
discord.py==1.7.3
